updates to generateTrip input schema.
